AWS Docker container
Write the Dockerfile for a container that will do the following tasks: (For creating dockerfiles, here is the link to the slides: https://container.training/intro-selfpaced.yml.html#176 Links to an external site.)

it will be based on Ubuntu
you will add a program (in any language) or shell script that will ask the user for their BlazerID and once the user enters the name, display the name using Figlet.
(hint: you need to start from the Ubuntu image, install figlet, copy your source code that (i) asks the user for the user's name, and (ii) take the entered name string and invokes figlet with the string passed as a parameter to figlet.)
Save the Dockerfile in your Docker instance. Run the Dockerfile, enter your name, and show the output from the instance.